feat: jest implementation w/ integration testing for 'users'

+++ src/controllers/user-controllers.ts
@@ -36,7 +36,7 @@ export async function register (req: Request, res: Response) {
 
     try {
         await userServices.createUser({ username, email, password });
-        res.status(httpStatus.CREATED).send({})
+        res.status(httpStatus.CREATED).send({ username, email })
     } catch (err: unknown) {
         const error = err as ApplicationError;
         errorHandler(error, req, res);
@@ -46,10 +46,10 @@ export async function register (req: Request, res: Response) {
 async function login (req: Request, res: Response) {
     const login = req.body as LoginUser;
     const { error } = loginSchema.validate(login);
-
     if (error) return res.status(httpStatus.BAD_REQUEST).send({ message: error.message });
 
     const { email, password } = req.body;
+
     try {
         const token = await userServices.loginUser({ email, password });
         if (token) {
@@ -67,10 +67,21 @@ export async function updateUser(req: Request, res: Response) {
     const userId = parseInt(id);
     const { username, email, password } = req.body;
 
+    const { error } = registerSchema.validate(req.body)
+    if (error) return res.status(httpStatus.BAD_REQUEST).send({ message: error.message });
+
+    const userToken = res.locals.user;
     try {
+        const user = await userServices.retrieveUserById(id);
+        if (!user) return res.status(httpStatus.BAD_REQUEST).send("User not found");
+
+        const user_id = await userServices.retrieveSession(userToken);
+        if (user.id !== user_id.user_id) return res.status(httpStatus.UNAUTHORIZED).send("Invalid request");
+
         const result = await userServices.updateUser({ username, email, password, userId })
         return res.status(httpStatus.OK).send({ result });
     } catch (err) {
+        console.log(err)
         const error = err as ApplicationError | Error;
         errorHandler(error, req, res);
     }
@@ -80,9 +91,16 @@ export async function deleteUser(req: Request, res: Response) {
     const { id } = req.params;
     const userId = parseInt(id);
 
+    const userToken = res.locals.user;
     try {
+        const user = await userServices.retrieveUserById(id);
+        if (!user) return res.status(httpStatus.BAD_REQUEST).send("User not found");
+
+        const user_id = await userServices.retrieveSession(userToken);
+        if (user.id !== user_id.user_id) return res.status(httpStatus.UNAUTHORIZED).send("Invalid request");
+
         await userServices.deleteUser({ userId })
-        return res.status(httpStatus.OK).send({});
+        return res.status(httpStatus.NO_CONTENT).send({});
     } catch (err) {
         const error = err as ApplicationError | Error;
         errorHandler(error, req, res)
